Introduce `spring-rabbitmq-client` for RabbitMQ AMQP 1.0
Fixes: https://github.com/spring-projects/spring-amqp/issues/2991
Fixes: https://github.com/spring-projects/spring-amqp/issues/2992
Fixes: https://github.com/spring-projects/spring-amqp/issues/2993

* Add `QueueInformation.type` since RabbitMQ AMQP Client exposes such an info

+++ spring-rabbit/src/test/java/org/springframework/amqp/rabbit/core/RabbitAdminTests.java
@@ -163,7 +163,7 @@ public class RabbitAdminTests extends NeedsManagementTests {
 		}
 	}
 
-	private int messageCount(RabbitAdmin rabbitAdmin, String queueName) {
+	private long messageCount(RabbitAdmin rabbitAdmin, String queueName) {
 		QueueInformation info = rabbitAdmin.getQueueInfo(queueName);
 		assertThat(info).isNotNull();
 		return info.getMessageCount();
